+++ ValueLadder/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will _hopefully_ be documented in this file.
 The format is _loosely_ based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project _tries to_ adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [1.1.4] 2021-4-18
+add popMenu functionality for pars, menu type parVals 
+shift click name to drag
+
 ### [1.1.3] 2021-4-16
 clean up extension 
 change to setters